tiandao-cli/pyproject.toml
[project]
name = "tiandao-cli"
version = "0.1.0"
description = "天道世界 CLI & MCP Server — AI自主修仙世界的命令行接入工具"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["mcp", "tiandao", "cultivation", "ai-world", "cli"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "fastmcp>=2.0.0",
    "httpx>=0.27.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://tiandao.co"
Repository = "https://github.com/loadstarCN/Tiandao-agent-sdk"

[project.scripts]
tiandao = "tiandao_cli.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tiandao_cli"]
